requirements.txt
PyPDF2
camelot-py
opencv-python
requests
pandas
bs4
tabula-py
selenium